+++ lib/pueri.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 require 'pueri/age'
+require 'pueri/dosecalc'
+require 'pueri/dosecheck'
 require 'pueri/neuro'
 require 'pueri/vax'
 require 'pueri/version'