+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * [Disclaimer](ABOUT.md), regarding the objectives and limitations of this project.
+* `DoseCalc`, a class for calculating the dosage of medications for pediatric patients.
+* `DoseCheck`, a class for checking the dosage-per-weight-day of prescribed medications.
 
 ### Changed
 * Bumped [RuboCop](https://github.com/rubocop-hq/rubocop) to `v.0.76.0`.